Server/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
cloudinary==1.41.0
exceptiongroup==1.2.2
Faker==27.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.32
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.20.0
gunicorn==22.0.0
psycopg2-binary==2.9.7
